refactor: clean up jss related hacks
