+++ src/components/InjectedComponents/ShadowRootSwitchNotifier.tsx
@@ -1,6 +1,6 @@
 import { sideEffect } from '../../utils/side-effects'
 import { renderInShadowRootSettings } from '../../settings/settings'
-import { renderInShadowRoot } from '../../utils/jss/renderInShadowRoot'
+import { renderInShadowRoot } from '../../utils/shadow-root/renderInShadowRoot'
 import { DOMProxy, GetContext } from '@holoflows/kit/es'
 import React from 'react'
 import { useSnackbar } from 'notistack'
