+++ src/plugins/RedPacket/UI/RedPacketDialog.tsx
@@ -29,8 +29,8 @@ import { useCapturedInput } from '../../../utils/hooks/useCapturedEvents'
 import { PluginMessageCenter } from '../../PluginMessages'
 import { getActivatedUI } from '../../../social-network/ui'
 import { formatBalance } from '../../Wallet/formatter'
-import ShadowRootDialog from '../../../utils/jss/ShadowRootDialog'
-import { PortalShadowRoot } from '../../../utils/jss/ShadowRootPortal'
+import ShadowRootDialog from '../../../utils/shadow-root/ShadowRootDialog'
+import { PortalShadowRoot } from '../../../utils/shadow-root/ShadowRootPortal'
 import BigNumber from 'bignumber.js'
 import { useSelectWallet, useTokens, useWallets } from '../../shared/useWallet'
 import { WalletSelect } from '../../shared/WalletSelect'
