+++ src/social-network-provider/twitter.com/ui/injectKnownIdentity.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { MutationObserverWatcher, ValueRef } from '@holoflows/kit'
 import { PersonKnown, PersonKnownProps } from '../../../components/InjectedComponents/PersonKnown'
 import { bioCardSelector } from '../utils/selector'
-import { renderInShadowRoot } from '../../../utils/jss/renderInShadowRoot'
+import { renderInShadowRoot } from '../../../utils/shadow-root/renderInShadowRoot'
 import { twitterUrl } from '../utils/url'
 import { makeStyles } from '@material-ui/core/styles'
 import { bioCardParser } from '../utils/fetch'
