+++ src/social-network-provider/twitter.com/ui/injectMaskbookIcon.tsx
@@ -5,7 +5,7 @@ import type { PostInfo } from '../../../social-network/PostInfo'
 import Services from '../../../extension/service'
 import { ProfileIdentifier } from '../../../database/type'
 import { MaskbookIcon } from '../../../resources/Maskbook-Circle-WhiteGraph-BlueBackground'
-import { renderInShadowRoot } from '../../../utils/jss/renderInShadowRoot'
+import { renderInShadowRoot } from '../../../utils/shadow-root/renderInShadowRoot'
 import { memoizePromise } from '../../../utils/memoize'
 import { Flags } from '../../../utils/flags'
