+++ src/social-network-provider/twitter.com/ui/injectPostDialogHint.tsx
@@ -3,7 +3,7 @@ import { useCallback } from 'react'
 import { twitterUrl } from '../utils/url'
 import { MutationObserverWatcher, LiveSelector } from '@holoflows/kit/es'
 import { postEditorInTimelineSelector, postEditorInPopupSelector } from '../utils/selector'
-import { renderInShadowRoot } from '../../../utils/jss/renderInShadowRoot'
+import { renderInShadowRoot } from '../../../utils/shadow-root/renderInShadowRoot'
 import { PostDialogHint } from '../../../components/InjectedComponents/PostDialogHint'
 import { makeStyles, Theme } from '@material-ui/core'
 import { MessageCenter } from '../../../utils/messages'
