+++ src/social-network-provider/twitter.com/ui/injectPostDialogIcon.tsx
@@ -3,7 +3,7 @@ import { useCallback } from 'react'
 import { twitterUrl } from '../utils/url'
 import { MutationObserverWatcher, LiveSelector } from '@holoflows/kit/es'
 import { posteditorToolbarSeelctor } from '../utils/selector'
-import { renderInShadowRoot } from '../../../utils/jss/renderInShadowRoot'
+import { renderInShadowRoot } from '../../../utils/shadow-root/renderInShadowRoot'
 import { PostDialogIcon } from '../../../components/InjectedComponents/PostDialogIcon'
 import { MessageCenter } from '../../../utils/messages'
 import { isCompose, isMobile } from '../utils/postBox'
