+++ src/social-network/defaults/injectCommentBox.tsx
@@ -3,7 +3,7 @@ import type { PostInfo } from '../PostInfo'
 import { MutationObserverWatcher } from '@holoflows/kit/es'
 import { CommentBox, CommentBoxProps } from '../../components/InjectedComponents/CommentBox'
 import Services from '../../extension/service'
-import { renderInShadowRoot } from '../../utils/jss/renderInShadowRoot'
+import { renderInShadowRoot } from '../../utils/shadow-root/renderInShadowRoot'
 import { dispatchCustomEvents, nop, selectElementContents, sleep } from '../../utils/utils'
 import { makeStyles } from '@material-ui/core'
 import { PostInfoContext, usePostInfoDetails, usePostInfo } from '../../components/DataSource/usePostInfo'
