+++ src/social-network/defaults/injectPageInspector.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { makeStyles } from '@material-ui/core'
 import { PageInspector, PageInspectorProps } from '../../components/InjectedComponents/PageInspector'
-import { renderInShadowRoot } from '../../utils/jss/renderInShadowRoot'
+import { renderInShadowRoot } from '../../utils/shadow-root/renderInShadowRoot'
 import { MutationObserverWatcher, LiveSelector } from '@holoflows/kit/es'
 import { Flags } from '../../utils/flags'
