+++ src/social-network/defaults/injectPostInspector.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import type { DOMProxy } from '@holoflows/kit'
 import type { PostInfo } from '../PostInfo'
-import { renderInShadowRoot } from '../../utils/jss/renderInShadowRoot'
+import { renderInShadowRoot } from '../../utils/shadow-root/renderInShadowRoot'
 import { PostInspector, PostInspectorProps } from '../../components/InjectedComponents/PostInspector'
 import { makeStyles } from '@material-ui/core'
 import { PostInfoContext } from '../../components/DataSource/usePostInfo'
