+++ src/social-network/defaults/injectPostReplacer.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { renderInShadowRoot } from '../../utils/jss/renderInShadowRoot'
+import { renderInShadowRoot } from '../../utils/shadow-root/renderInShadowRoot'
 import { PostInfoContext } from '../../components/DataSource/usePostInfo'
 import { PostReplacer, PostReplacerProps } from '../../components/InjectedComponents/PostReplacer'
 import type { PostInfo } from '../PostInfo'
