+++ src/utils/shadow-root/ShadowRootDialog.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Dialog, withMobileDialog, DialogProps } from '@material-ui/core'
-import '../../utils/jss/ShadowRootPortal'
-import { PortalShadowRoot } from '../../utils/jss/ShadowRootPortal'
+import '../../utils/shadow-root/ShadowRootPortal'
+import { PortalShadowRoot } from '../../utils/shadow-root/ShadowRootPortal'
 import { useSheetsRegistryStyles } from './renderInShadowRoot'
 import { useValueRef } from '../hooks/useValueRef'
 import { renderInShadowRootSettings } from '../../settings/settings'
